Check if role is ADMIN then can not be deleted

+++ backend/controller/deleteUser.js
@@ -4,12 +4,27 @@ async function deleteUser(req, res) {
   try {
     let query = { _id: req.body.userId.toString() };
 
-    await userModel.deleteOne(query);
-    res.json({
-      message: "User Deleted",
-      success: true,
-      error: false,
-    });
+    await userModel.findById(query)
+    .then(user => {
+      if(user.role === "ADMIN") {
+        res.json({
+          message: "How dare you delete your boss?",
+          success: false,
+          error: true,
+        });
+      }
+      else {
+        userModel.deleteOne(query);
+        res.json({
+          message: "Deleted User",
+          success: true,
+          error: false,
+        });
+      }
+    })
+
+    
+    
   } catch (err) {
     res.status(400).json({
       message: err.message || err,
